requirements.txt
nvidia-ml-py3>=12.535.108
numpy>=1.21.0
psutil>=5.8.0
pyyaml>=6.0